requirements.txt
django-environ
python-decouple
django-ace
requests
markdown
tiktoken
openai